requirements.txt
fastapi
uvicorn
requests
python-dotenv
elasticsearch
google-genai
aiohttp